% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{aux_addNamesToPercentiles}
\alias{aux_addNamesToPercentiles}
\title{Add names to percentiles vector}
\usage{
aux_addNamesToPercentiles(percentiles)
}
\arguments{
\item{percentiles}{Vector with percentiles values.}
}
\value{
The vector \code{percentiles} with names.
}
\description{
Add names to percentiles vector
}
\seealso{
Other Data: 
\code{\link{aux_constructCIpercentiles}()},
\code{\link{remove_duplicates}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Data}
