% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{aux_CommonSubPath}
\alias{aux_CommonSubPath}
\title{Common Sub-Path}
\usage{
aux_CommonSubPath(x)
}
\arguments{
\item{x}{Vector of paths for which the sub-path needs to be detected}
}
\value{
Character of the common sub-path
}
\description{
Identify the common sub-path in the vector of paths \code{x}
}
\examples{
Paths <- c("C:/Users/Default",
           "C:/Users/Public")
aux_CommonSubPath(Paths)

}
\seealso{
Other File: 
\code{\link{file.copyMMV}()},
\code{\link{get_fileNameExtension}()},
\code{\link{is.fileMMV}()},
\code{\link{load_RDataAsList}()},
\code{\link{saveMMV}()},
\code{\link{thisFile}()}
}
\author{
Anne Kuemmel (IntiQuan), Nathalie Gobeau (MMV, \email{gobeaun@mmv.org})
}
\concept{File}
