% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{aux_constructCIpercentiles}
\alias{aux_constructCIpercentiles}
\title{Construct CI percentiles from CI level}
\usage{
aux_constructCIpercentiles(CIlevel)
}
\arguments{
\item{CIlevel}{Numeric containing the confidence interval in percent to be estimated (Default: 90)}
}
\value{
The vector \code{CI.percentiles} with names.
}
\description{
Construct CI percentiles from CI level
}
\seealso{
Other Data: 
\code{\link{aux_addNamesToPercentiles}()},
\code{\link{remove_duplicates}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Data}
