% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{aux_createUSUBJID}
\alias{aux_createUSUBJID}
\title{Create USUBJID}
\usage{
aux_createUSUBJID(data)
}
\arguments{
\item{data}{A data frame containing the columns \code{STUDY}, \code{GROUP}, and \code{SUBJECT}.}
}
\value{
A character vector that can be assigned as USUBJID column to data.
}
\description{
Add column \code{USUBJID} to \code{data} based on columns \code{STUDY}, \code{GROUP}, and \code{SUBJECT},
when preparing IQR data from SCID dataset.
}
\details{
The function creates a unique subject identifier (\code{USUBJID}) by concatenating the values
of the \code{STUDY}, \code{GROUP}, and \code{SUBJECT} columns. The \code{GROUP} and \code{SUBJECT} values
are formatted to have at least two digits.
}
\examples{
data <- data.frame(STUDY   = "StudyID",
                   GROUP   = c("G1","G1","G2","G2"),
                   SUBJECT = c("M1","M2", "M1", "M2"),
                   stringsAsFactors = FALSE)
data$USUBJID <- aux_createUSUBJID(data)

}
\seealso{
Other Data: 
\code{\link{Check_MissingDatabyNAME}()},
\code{\link{IQRtableToDataFrame}()},
\code{\link{aux_constructCIpercentiles}()},
\code{\link{check_dataGeneralMMV_Center}()},
\code{\link{check_dataGeneralMMV}()},
\code{\link{convert_Unit}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{summarize_PIandCIgeneric}()},
\code{\link{summaryAcrossTrials}()},
\code{\link{summaryByTrial}()},
\code{\link{swapName_MMVnameToName}()},
\code{\link{swapName_NameToMMVname}()},
\code{\link{transform_dataFrame_LongToWide}()},
\code{\link{transform_dataFrame_WideToLong}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan)
}
\concept{Data}
