% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{aux_formatErrorName}
\alias{aux_formatErrorName}
\title{Format Error Name}
\usage{
aux_formatErrorName(parameterNames)
}
\arguments{
\item{parameterNames}{Vector of character with the names of the parameters}
}
\value{
\code{parameterNames} with converted format
}
\description{
Converts Error Parameter Names as used in sysFIT to format as used in NLME
}
\seealso{
Other Utility: 
\code{\link{aux_addNamesToPercentiles}()},
\code{\link{aux_removeEscapeChar}()},
\code{\link{collapseMMV}()},
\code{\link{get_ActivityPath}()},
\code{\link{remove_duplicates}()}
}
\author{
Anne Kuemmel (IntiQuan)
}
\concept{Utility}
