% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{aux_removeEscapeChar}
\alias{aux_removeEscapeChar}
\title{Removal of escape characters}
\usage{
aux_removeEscapeChar(x, escapeChars = c("\\r\\n", "\\n", "\\r", "\\t"))
}
\arguments{
\item{x}{A character vector.}

\item{escapeChars}{Escape character to substituted with space (Default: \code{c("\\r\\n", "\\n", "\\r", "\\t")})}
}
\value{
\code{x} where all escape character defined in \code{escapeChars} were replaced by a space.
}
\description{
Removal of escape characters
}
\seealso{
Other Utility: 
\code{\link{aux_addNamesToPercentiles}()},
\code{\link{aux_formatErrorName}()},
\code{\link{collapseMMV}()},
\code{\link{get_ActivityPath}()},
\code{\link{remove_duplicates}()}
}
\author{
Anne Kuemmel (IntiQuan)
}
\concept{Utility}
