% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Check_MissingDatabyNAME}
\alias{Check_MissingDatabyNAME}
\title{Check_MissingDatabyNAME}
\usage{
Check_MissingDatabyNAME(dataMaster, NAME, colNAME = c("USUBJID"))
}
\arguments{
\item{dataMaster}{Master dataset in MMV format}

\item{NAME}{list of variable to test}

\item{colNAME}{Column name to use to identify subject (default: "USUBJID")}
}
\value{
dataSummaryMissing, which is a data.frame with the list of missing subject for each variable in \code{NAME}
}
\description{
Check for for each variable defined in \code{NAME}, if it is missing for
any subject defined in the column USUBJID  of the dataset \code{dataMaster}.
}
\seealso{
Other Data Preparation: 
\code{\link{check_dataGeneralMMV}()},
\code{\link{check_dataGeneralMMV_Center}()},
\code{\link{swapName_MMVnameToName}()},
\code{\link{swapName_NameToMMVname}()},
\code{\link{transform_dataFrame_LongToWide}()},
\code{\link{transform_dataFrame_WideToLong}()}
}
\author{
Mohammed H. Cherkaoui (MMV), \email{cherkaouim@mmv.org}
}
\concept{Data Preparation}
