% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conversionTable}
\alias{conversionTable}
\title{Conversion Table}
\format{
A data frame with several rows and columns:
\describe{
\item{UNIT_IN}{The unit to convert from.}
\item{UNIT_OUT}{The unit to convert to.}
\item{FORMULA}{The formula to use for the conversion, e.g. X_out=X_in*1000.}
\item{COMMENT}{A comment on the conversion.}
}
}
\source{
Internal MMV data
}
\usage{
conversionTable
}
\description{
A dataset containing conversion factors for various units.
}
\details{
This dataset is used to convert between different units of measurement.
}
\examples{
data(conversionTable)
head(conversionTable)

}
\keyword{datasets}
