% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_Unit.R
\name{convert_Unit}
\alias{convert_Unit}
\title{Convert Unit}
\usage{
convert_Unit(
  data,
  UNIT_OUT,
  colUNIT = "UNIT",
  colVALUE = "VALUE",
  conversionTable = NULL
)
}
\arguments{
\item{data}{data.frame where to convert value}

\item{UNIT_OUT}{Desired unit for output}

\item{colUNIT}{Name of the column unit (Default: \code{"UNIT"})}

\item{colVALUE}{Name of the column value (Default: \code{"VALUE"})}

\item{conversionTable}{Conversion table - a data.frame with the following character columns:
\itemize{
\item UNIT_IN
\item UNIT_OUT
\item FORMULA
\item COMMENT
If \code{NULL} (default), the internal conversion table of the MMVbase
package will be used.
}}
}
\value{
data.frame identical to \code{data} but with the \code{colVALUE} adjust to the desired output unit as defined in \code{UNIT_OUT}.
}
\description{
Automate the conversion of unit in a data frame.
}
\examples{
dat <- data.frame(TIME  = c( 0, 1, 2, 4, 8),
                  VALUE = c(10, 6, 4, 3, 2),
                  UNIT  = "ng/mL",
                  stringsAsFactors = FALSE)
 dat <- convert_Unit(data     = dat,
                     UNIT_OUT = c("ug/mL"))

}
\seealso{
Other General Functions: 
\code{\link{IQRtableToDataFrame}()},
\code{\link{aux_CommonSubPath}()},
\code{\link{aux_createUSUBJID}()},
\code{\link{aux_formatErrorName}()},
\code{\link{aux_removeEscapeChar}()},
\code{\link{clopperPearsonMMV}()},
\code{\link{dichotomicSearch}()},
\code{\link{file.copyMMV}()},
\code{\link{find_MinMMV}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rectintMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{thisFile}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
