% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTable.R
\name{CreateOrAppendToReportTable}
\alias{CreateOrAppendToReportTable}
\title{Generate a ReportTable or add rows to an existing one.}
\usage{
CreateOrAppendToReportTable(
  table = NULL,
  Style = NA_character_,
  Text = NA_character_,
  ...
)
}
\arguments{
\item{table}{a data.frame or NULL (default: NULL).}

\item{Style}{a character string. Default: NA_character_.}

\item{Text}{a character string. Default: NA_character_.}

\item{...}{name value pairs to become columns in the table.}
}
\value{
a data.frame
}
\description{
Generate a ReportTable or add rows to an existing one.
}
\examples{

# Code to put/adapt in the beginning of your script or function definition:
envResultTable <- new.env()
envResultTable$resultTable <- NULL
AppendResult <- function(...) {
  envResultTable$resultTable <- CreateOrAppendToReportTable(
    envResultTable$resultTable, ...)
}
AppendParameter <- function(
  Parameter, 
  Rat = NA_real_, Mouse = NA_real_, Dog = NA_real_, Human = NA_real_, 
  Value = NA_real_, Unit = NA_character_) {
  
AppendResult(
  Parameter = Parameter, 
  Rat = list(Rat), Mouse = list(Mouse), Dog = list(Dog), Human = list(Human), 
  Value = list(Value), Unit = Unit)
}

# Results added "on the go" in the implementation of a function or a procedure 
# in a script
AppendResult(Style = "subtitle", Text = "PK Parameter Summary (plasma based)")

# Parameter values can be both, character or numeric
AppendParameter("Selected Prediction Method", Value = "IVIVE", Unit = "method")

CL_plasma_mLminkg <- 20
AppendParameter("Total Clearance", Human = CL_plasma_mLminkg, Unit = "mL/min/kg")
RenalClearance_H_mLminkg <- 3
AppendResult(
  Style = "comment", 
  Text = "Hepatic clearance obtained by substracting renal from total plasma clearanc")
AppendParameter(
  "Hepatic Clearance", 
  Human = CL_plasma_mLminkg-RenalClearance_H_mLminkg, Unit = "mL/min/kg")

# At the end of the function return the accumulated result table or assign it 
# to a variable to use later:
resultTable <- envResultTable$resultTable

# Getting a predicted parameter value from the result table
Human_CL_plasma_mLminkg <- get_FromTable(
  resultTable, "Hepatic Clearance", "Human")
}
\author{
Venelin Mitov
}
