% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.copyMMV}
\alias{file.copyMMV}
\title{Copy a file with creating the destination folder if it does not exist}
\usage{
file.copyMMV(
  from,
  to,
  overwrite = recursive,
  recursive = FALSE,
  copy.mode = TRUE,
  copy.date = FALSE
)
}
\arguments{
\item{from}{Character vectors, containing file names or paths.}

\item{to}{Character vectors, containing file names or paths.}

\item{overwrite}{Logical; should existing destination files be overwritten?}

\item{recursive}{Logical. If to is a directory, should directories in from be copied (and their contents)? (Like cp -R on POSIX OSes.)}

\item{copy.mode}{Logical: should file permission bits be copied where possible?}

\item{copy.date}{Logical: should file dates be preserved where possible? See Sys.setFileTime.}
}
\description{
\link{file.copyMMV} works in a similar way to \link[base:files]{base::file.copy} but automatically create
the destination folder if needed. Copying to existing destination files is skipped
unless \code{overwrite = TRUE}. The to argument can specify a single existing directory.
If \code{copy.mode = TRUE} file read/write/execute permissions are copied where possible,
restricted by \code{'umask'}. (On Windows this applies only to files.) Other security
attributes such as ACLs are not copied. On a POSIX filesystem the targets of symbolic
links will be copied rather than the links themselves, and hard links are copied separately.
Using \code{copy.date = TRUE} may or may not copy the timestamp exactly (for example, fractional
seconds may be omitted), but is more likely to do so as from R 3.4.0.
}
\seealso{
\link[base:files]{base::file.copy}

Other File: 
\code{\link{aux_CommonSubPath}()},
\code{\link{get_fileNameExtension}()},
\code{\link{is.fileMMV}()},
\code{\link{load_RDataAsList}()},
\code{\link{saveMMV}()},
\code{\link{thisFile}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{File}
