% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical.R
\name{find_MinMMV}
\alias{find_MinMMV}
\title{Find minimum of a continuous function (specified by x,y values) by interpolation}
\usage{
find_MinMMV(X, Y, dx = as.numeric((X[2] - X[1])/1000), Method = "Quadratic")
}
\arguments{
\item{X}{numeric vector}

\item{Y}{numeric vector, same length as X}

\item{dx}{numeric value, refined interpolation interval in case of the cubic spline interpolation}

\item{Method}{interpolation method to choose from  \code{"NoInterpolation"}, \code{"CubicSpline"} or \code{"Quadratic"} (Default: \code{"Quadratic"})}
}
\value{
Data frame with \code{Xmin} anf \code{Ymin}: \code{data.frame(Xmin=Xmin,Ymin=Ymin)}
}
\description{
Finds the minimum of \code{Y(X)} by interpolation with one of the 3 methods: no interpolation, cubic spline or quadratic.
Returns the minimum value of \code{Y} (\code{Ymin}) and the value of \code{X} when \code{Y} is minimum (\code{Xmin})
}
\examples{
X  = c(1,2,3,4,5,6,7)
Y  = c(10,4,1,1,6,10,20)
Y1 = find_MinMMV(X,Y,Method = "NoInterpolation")
Y2 = find_MinMMV(X,Y,Method = "CubicSpline")
Y3 = find_MinMMV(X,Y,Method = "Quadratic")
plot(X,Y,type="b",ylim=c(0,20))
points(Y1,pch=16,col="blue")
points(Y2,pch=16,col="green")
points(Y3,pch=16,col="red")
legend("topleft",legend=c("NoInterpolation","CubicSpline","Quadratic"),
       pch=rep(16,3),col=c("blue","green","red"))

}
\seealso{
Other Mathematical: 
\code{\link{IsInClosedInterval}()},
\code{\link{IsInLeftOpenInterval}()},
\code{\link{IsInOpenInterval}()},
\code{\link{IsInRightOpenInterval}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rectintMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Mathematical}
