% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_CLhModels.R
\name{formula_CLhPlToCLintu}
\alias{formula_CLhPlToCLintu}
\title{Estimate intrinsic unbound hepatic clearance from final hepatic plasma clearance}
\usage{
formula_CLhPlToCLintu(CLh, Qh, Fu, BP)
}
\arguments{
\item{CLh}{numeric vector indicating hepatic plasma CL (mL/min/kg).}

\item{Qh}{numeric vector indicating hepatic blood flow (mL/min/kg).}

\item{Fu}{numeric vector indicating fraction unbound in plasma (.).}

\item{BP}{numeric vector indicating BP-ratio (.), used to convert plasma to blood matrix for
CLh and Fu.}
}
\value{
Estimated intrinsic unbound hepatic clearance (mL/min/kg) or an error
if the arguments don't correspond.
}
\description{
Estimate intrinsic unbound hepatic clearance from final hepatic plasma clearance
}
\details{
Currently using the well-stirred model. If arguments are vectors
they should be the same length as CLh, otherwise, an error is thrown.
}
\seealso{
formula_CLhToCLintu

Other Hepatic CL: 
\code{\link{formula_CLhToCLintu}()},
\code{\link{formula_CLintuToCLh}()},
\code{\link{formula_CLintuToCLhPl}()},
\code{\link{formula_CLintuToCLh_Parallel}()},
\code{\link{formula_CLintuToCLh_WellStirred}()},
\code{\link{formula_HepExtraction}()}
}
\author{
Venelin Mitov (IntiQuan)
}
\concept{Hepatic CL}
