% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_CLhModels.R
\name{formula_CLhToCLintu}
\alias{formula_CLhToCLintu}
\title{Convert hepatic blood clearance to intrinsic clearance (unbound)}
\usage{
formula_CLhToCLintu(CLh, Qh, fub, Model = "Well-Stirred")
}
\arguments{
\item{CLh}{Numeric vector - hepatic clearance (mL/min, L/hr or L/hr/kg).}

\item{Qh}{numeric vector - hepatic blood flow (mL/min, L/hr or L/hr/kg); vector length,
names and units should be consistent with CLh.}

\item{fub}{numeric vector - fraction unbound in Blood (.). Vector length, names and units should be
consistent with CLh.}

\item{Model}{Character string indicating the model to use for conversion (
possible options: "Well-Stirred", "Parallel").}
}
\value{
numeric vector of the same length as CLh - intrinsic clearance (unbound) (mL/min, L/hr or L/hr/kg).
}
\description{
This function converts hepatic blood clearance (CLh) to intrinsic clearance (CLintu) using different models.
}
\details{
The function converts hepatic clearance (CLh) to intrinsic clearance (CLintu) using different models.
The available models are "Well-Stirred" and "Parallel".
If the model is "Well-Stirred", the function uses the formula_CLhToCLintu_WellStirred() function.
If the model is "Parallel", the function uses the formula_CLhToCLintu_Parallel() function.
If the model is not one of the available options, an error is thrown.
}
\examples{
# Convert hepatic clearance to intrinsic clearance using the Well-Stirred model
formula_CLhToCLintu(CLh = 100, Qh = 200, fub = 0.5, Model = "Well-Stirred")

# Convert hepatic clearance to intrinsic clearance using the Parallel model
formula_CLhToCLintu(CLh = 100, Qh = 200, fub = 0.5, Model = "Parallel")

}
\seealso{
\code{\link{formula_CLhToCLintu_WellStirred}}, \code{\link{formula_CLhToCLintu_Parallel}}

Other Hepatic CL: 
\code{\link{formula_CLhPlToCLintu}()},
\code{\link{formula_CLintuToCLh}()},
\code{\link{formula_CLintuToCLhPl}()},
\code{\link{formula_CLintuToCLh_Parallel}()},
\code{\link{formula_CLintuToCLh_WellStirred}()},
\code{\link{formula_HepExtraction}()}
}
\author{
Mohammed H. Cherkaoui (MMV), Venelin Mitov (IntiQuan)
}
\concept{Hepatic CL}
