% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_CLhModels.R
\name{formula_CLintuToCLh}
\alias{formula_CLintuToCLh}
\title{Convert unbound intrinsic clearance to hepatic blood clearance}
\usage{
formula_CLintuToCLh(CLintu, Qh, fub, Model = "Well-Stirred")
}
\arguments{
\item{CLintu}{numeric vector - unbound intrinsic clearance (mL/min, L/hr or L/hr/kg).}

\item{Qh}{Hepatic blood flow (mL/min, L/hr or L/hr/kg) - Should be consistent with CLintu}

\item{fub}{Fraction unbound in Blood (.)}

\item{Model}{Model to use for conversion. Options are "Well-Stirred" or "Parallel".}
}
\value{
Hepatic  blood clearance (CLh) in the same units as CLintu and Qh.
}
\description{
This function converts the unbound intrinsic clearance (CLintu) to hepatic
clearance (CLh) using either the Well-Stirred or Parallel model.
}
\examples{
# Example 1: Using the Well-Stirred model
formula_CLintuToCLh(CLintu = 10, Qh = 100, fub = 0.2, Model = "Well-Stirred")

# Example 2: Using the Parallel model
formula_CLintuToCLh(CLintu = 10, Qh = 100, fub = 0.2, Model = "Parallel")

}
\seealso{
\code{\link{formula_CLintuToCLh_WellStirred}}, \code{\link{formula_CLintuToCLh_Parallel}}

Other Hepatic CL: 
\code{\link{formula_CLhPlToCLintu}()},
\code{\link{formula_CLhToCLintu}()},
\code{\link{formula_CLintuToCLhPl}()},
\code{\link{formula_CLintuToCLh_Parallel}()},
\code{\link{formula_CLintuToCLh_WellStirred}()},
\code{\link{formula_HepExtraction}()}
}
\author{
Mohammed H. Cherkaoui (MMV), Venelin Mitov (IntiQuan)
}
\concept{Hepatic CL}
