% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_CLhModels.R
\name{formula_CLintuToCLh_Parallel}
\alias{formula_CLintuToCLh_Parallel}
\alias{formula_CLhToCLintu_Parallel}
\title{Calculate hepatic blood clearance (CLh) from unbound intrinsic clearance (CLintu) or vice versa using the Parallel-Tube model.}
\usage{
formula_CLintuToCLh_Parallel(CLintu, Qh, fub)

formula_CLhToCLintu_Parallel(CLh, Qh, fub)
}
\arguments{
\item{CLintu}{Numeric, unbound intrinsic clearance (mL/min, L/hr or L/hr/kg).}

\item{Qh}{Numeric, blood flow rate through the liver (mL/min, L/hr or L/hr/kg).
The unit should be consistent with CLintu and/or CLh.}

\item{fub}{Numeric, fraction of drug unbound in the blood.}

\item{CLh}{Numeric, hepatic blood clearance (mL/min, L/hr or L/hr/kg).}
}
\value{
numeric, for formula_CLintuToCLh_Parallel - the hepatic blood clearance
CLh (mL/min, L/hr or L/hr/kg), for formula_CLhToCLintu_Parallel - the
unbound intrinsic clearance CLintu (mL/min, L/hr or L/hr/kg).
}
\description{
The Parallel-Tube model assumes that the liver is composed
of multiple parallel tubes, each with the same flow rate and drug concentration. It takes into
account the blood flow rate through the liver (Qh), the fraction of drug unbound in the blood (fub),
and the unbound intrinsic clearance (CLintu).
The equation used to calculate CLh from CLintu is
CLh = Qh * (1 - exp(-fub * CLintu / Qh)).
The equation used to calculate CLintu from CLh is
CLintu = -Qh / fub * log(1 - CLh / Qh).
}
\examples{
CLintu <- 10 # mL/min
Qh <- 1000 # mL/min
fub <- 0.5
CLh <- formula_CLintuToCLh_Parallel(CLintu, Qh, fub)
CLh
CLintu <- formula_CLhToCLintu_Parallel(CLh, Qh, fub)
CLintu

}
\references{
\enumerate{
\item Pang KS, Rowland M. Hepatic clearance of drugs. I. Theoretical considerations of a "well-stirred"
model and a "parallel tube" model. Influence of hepatic blood flow, plasma and blood cell binding,
and the hepatocellular enzymatic activity on hepatic drug clearance.
J Pharmacokinet Biopharm. 1977;5(6):625-653. doi.org/10.1007/BF01059688
}
}
\seealso{
Other Hepatic CL: 
\code{\link{formula_CLhPlToCLintu}()},
\code{\link{formula_CLhToCLintu}()},
\code{\link{formula_CLintuToCLh}()},
\code{\link{formula_CLintuToCLhPl}()},
\code{\link{formula_CLintuToCLh_WellStirred}()},
\code{\link{formula_HepExtraction}()}

Other Hepatic CL: 
\code{\link{formula_CLhPlToCLintu}()},
\code{\link{formula_CLhToCLintu}()},
\code{\link{formula_CLintuToCLh}()},
\code{\link{formula_CLintuToCLhPl}()},
\code{\link{formula_CLintuToCLh_WellStirred}()},
\code{\link{formula_HepExtraction}()}
}
\author{
Mohammed H. Cherkaoui (MMV), Venelin Mitov (IntiQuan)
}
\concept{Hepatic CL}
