% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_CLhModels.R
\name{formula_CLintuToCLh_WellStirred}
\alias{formula_CLintuToCLh_WellStirred}
\alias{formula_CLhToCLintu_WellStirred}
\title{Calculate hepatic blood clearance (CLh) from unbound intrinsic clearance (CLintu) vice versa using the Well-Stirred model}
\usage{
formula_CLintuToCLh_WellStirred(CLintu, Qh, fub)

formula_CLhToCLintu_WellStirred(CLh, Qh, fub)
}
\arguments{
\item{CLintu}{Numeric, unbound Intrinsic Clearance (mL/min, L/hr or L/hr/kg).}

\item{Qh}{Numeric, blood flow rate through the liver (mL/min, L/hr or L/hr/kg).
The unit should be consistent with CLint.}

\item{fub}{Numeric, fraction of drug unbound in the blood.}

\item{CLh}{Numeric, hepatic Clearance (mL/min, L/hr or L/hr/kg).}
}
\value{
The function formula_CLintuToCLh_WellStirred returns CLh -
Hepatic clearance (mL/min, L/hr or L/hr/kg). The function
formula_CLhToCLintu_WellStirred returns CLintu - unbound intrinsic clearance
(mL/min, L/hr or L/hr/kg).
}
\description{
The Well-Stirred model assumes that the liver is a
well-mixed organ and that drug elimination occurs primarily in the liver.
It takes into account the blood flow rate through the liver (Qh) and the
fraction of drug unbound in the blood (fub).
formula_CLintuToCLh_WellStirred
calculates hepatic clearance (CLh) from unbound intrinsic clearance (CLintu)
using the equation:
CLh = Qh * fub * CLintu / (Qh + fub * CLintu).
formula_CLhToCLintu_WellStirred
calculates unbound intrinsic clearance (CLintu) from hepatic clearance (CLh)
using the equation:
CLintu = 1 / fub * Qh * CLh / (Qh - CLh).
}
\examples{
CLintu <- 10 # mL/min
Qh <- 1000 # mL/min
fub <- 0.5
CLh <- formula_CLintuToCLh_WellStirred(CLintu, Qh, fub)
CLh
CLintu <- formula_CLhToCLintu_WellStirred(CLh, Qh, fub)
CLintu

}
\references{
\enumerate{
\item Pang KS, Rowland M. Hepatic clearance of drugs. I. Theoretical considerations of a "well-stirred"
model and a "parallel tube" model. Influence of hepatic blood flow, plasma and blood cell binding,
and the hepatocellular enzymatic activity on hepatic drug clearance.
J Pharmacokinet Biopharm. 1977;5(6):625-653. doi.org/10.1007/BF01059688
}
}
\seealso{
formula_CLintuToCLh_Parallel

Other Hepatic CL: 
\code{\link{formula_CLhPlToCLintu}()},
\code{\link{formula_CLhToCLintu}()},
\code{\link{formula_CLintuToCLh}()},
\code{\link{formula_CLintuToCLhPl}()},
\code{\link{formula_CLintuToCLh_Parallel}()},
\code{\link{formula_HepExtraction}()}

Other Hepatic CL: 
\code{\link{formula_CLhPlToCLintu}()},
\code{\link{formula_CLhToCLintu}()},
\code{\link{formula_CLintuToCLh}()},
\code{\link{formula_CLintuToCLhPl}()},
\code{\link{formula_CLintuToCLh_Parallel}()},
\code{\link{formula_HepExtraction}()}
}
\author{
Catalina Barcelo (MMV), Mohammed H. Cherkaoui (MMV), Venelin Mitov (IntiQuan)

Mohammed H. Cherkaoui (MMV), Venelin Mitov (IntiQuan)
}
\concept{Hepatic CL}
