% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_EMAXModel.R
\name{formula_EC50toMPC90}
\alias{formula_EC50toMPC90}
\alias{formula_EC50toEClevel}
\title{Convert EC50 to EC for a desired effect level in the EMAX model}
\usage{
formula_EC50toMPC90(EC50, hill)

formula_EC50toEClevel(EC50, hill, level)
}
\arguments{
\item{EC50}{numeric indicating concentration.}

\item{hill}{numeric indicating hill parameter value.}

\item{level}{a numeric in the interval \code{(0, 1)}}
}
\value{
formula_EC50toEClevel returns a numeric - the ECx corresponding to \code{level*100}\% of the maximum effect.
formula_EC50toMPC90 returns the MPC90 corresponding to EC50 and hill.
}
\description{
\itemize{
\item formula_EC50toEClevel: Calculate the ECx corresponding to a given level of the maximum effect.
\item formula_EC50toMPC90: Calculate the MPC90 from EC50 and Hill according to the EMAX model. MPC90
is the concentration at which 90\% of the maximum effect is reached.
}
}
\details{
The formula is derived from the EMAX model: EClevel = EC50*(level/(1-level))^(1/hill).
The function supports some or all of the input values being vectors.
}
\examples{
formula_EC50toEClevel(23, 3, 0.5)
formula_EC50toEClevel(23, 10, 0.9)
formula_EC50toEClevel(23, 1, 0.9)
}
\seealso{
Other EMAX: 
\code{\link{formula_EMAXEffect}()},
\code{\link{formula_EMAXmodelParsToMIC}()},
\code{\link{formula_PRRtoEMAX}()}
}
\author{
Anne Kuemmel (IntiQuan), Venelin Mitov (IntiQuan)
}
\concept{EMAX}
