% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_EMAXModel.R
\name{formula_EMAXEffect}
\alias{formula_EMAXEffect}
\title{Calculate the Effect of an EMAX model for a given concentration x ####}
\usage{
formula_EMAXEffect(x, EC50, hill)
}
\arguments{
\item{x}{numeric non-negative vector of concentration at which to calculate the effect.}

\item{EC50}{a positive numeric indicating the EC50 parameter value0}

\item{hill}{a positive numeric indicating the hill parameter value}
}
\value{
a numeric vector of the same length as x with values in the interval [0,1).
}
\description{
Calculate the Effect of an EMAX model for a given concentration x ####
}
\examples{
curve(formula_EMAXEffect(x, 23, 10), .1, 50)
}
\seealso{
Other EMAX: 
\code{\link{formula_EC50toMPC90}()},
\code{\link{formula_EMAXmodelParsToMIC}()},
\code{\link{formula_PRRtoEMAX}()}
}
\author{
Anne Kuemmel (IntiQuan), Venelin Mitov (IntiQuan)
}
\concept{EMAX}
