% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_EMAXModel.R
\name{formula_EMAXmodelParsToMIC}
\alias{formula_EMAXmodelParsToMIC}
\title{Minimum parasitemia growth inhibitory concentration (MIC) according to an EMAX PD model}
\usage{
formula_EMAXmodelParsToMIC(GR, EMAX, EC50, hill)
}
\arguments{
\item{GR}{positive numeric indicating the parasite growth rate (1/hours).}

\item{EMAX}{positive numeric indicating the maximum killing effect (1/hours).}

\item{EC50}{positive numeric indicating the concentration at which half of EMAX is reached.}

\item{hill}{positive numeric indicating the hill coefficient.}
}
\value{
numeric, MIC value. If EMAX <= GR, Inf is returned.
}
\description{
Determine MIC from EMAX PD model (sigmoidal emax model). MIC corresponds to
concentration at which the killing compensates growth. If EMAX <= GR, Inf is returned.
}
\details{
Not to confuse Minimum Inhibitory Concentration (MIC) with MICrosomes,
MIC will be noted MinIC.
}
\seealso{
Other EMAX: 
\code{\link{formula_EC50toMPC90}()},
\code{\link{formula_EMAXEffect}()},
\code{\link{formula_PRRtoEMAX}()}
}
\author{
Anne Kuemmel (IntiQuan), Venelin Mitov (IntiQuan)
}
\concept{EMAX}
