% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_CLhModels.R
\name{formula_HepExtraction}
\alias{formula_HepExtraction}
\title{Calculate hepatic extraction ratio from hepatic blood clearance and hepatic blood flow}
\usage{
formula_HepExtraction(CLh, Qh)
}
\arguments{
\item{CLh}{numeric - hepatic clearance.}

\item{Qh}{numeric - hepatic blood flow.}
}
\value{
numeric - hepatic extraction ratio.
}
\description{
Calculate hepatic extraction ratio from hepatic blood clearance and hepatic blood flow
}
\details{
The hepatic extraction ratio is equal to CLh/Qh.
}
\examples{
# hepatic clearance [mL/min/kg]
CLh_mLminkg <- 23
# Blood to plasma ratio
BPratio <- 2
# Hepatic blood flow [mL/min/kg]
Qh_mLminkg <- get_NumericConstantsAsList()$QhPatient_mLPerMinPerkg

Qh_mLminkg

# Hepatic extraction ratio
Eh <- formula_HepExtraction(CLh = CLh_mLminkg / BPratio, Qh = Qh_mLminkg)
# Hepatic bioavailability
Fh <- 1 - Eh

}
\seealso{
Other Hepatic CL: 
\code{\link{formula_CLhPlToCLintu}()},
\code{\link{formula_CLhToCLintu}()},
\code{\link{formula_CLintuToCLh}()},
\code{\link{formula_CLintuToCLhPl}()},
\code{\link{formula_CLintuToCLh_Parallel}()},
\code{\link{formula_CLintuToCLh_WellStirred}()}
}
\author{
Anne Kuemmel (IntiQuan), Venelin Mitov (IntiQuan)
}
\concept{Hepatic CL}
