% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_EMAXModel.R
\name{formula_PRRtoEMAX}
\alias{formula_PRRtoEMAX}
\alias{formula_EMAXtoPRR}
\title{Determine EMAX value from PRR or PRR value from EMAX accounting for parasite growth}
\usage{
formula_PRRtoEMAX(GR, PRR, timePRR = 48)

formula_EMAXtoPRR(GR, EMAX, timePRR = 48)
}
\arguments{
\item{GR}{Growth rate coefficient (1/hours)}

\item{PRR}{Parasite reduction ratio in log10 units}

\item{timePRR}{Hours over which PRR was assessed (default: 48 hours)}

\item{EMAX}{Maximum killing effect (1/hours).}
}
\value{
formula_PRRtoEMAX numeric - EMAX value (1/hours);
formula_EMAXtoPRR returns numeric - PRR value (log10 units).
}
\description{
Emax is the maximum killing effect. It is assumed that the PRR
was obtained over a period at which the drug achieved maximum
effect. The observed PRR is the difference of parasite growth
and killing by the drug. The derivation follows the EMAX model:
EMAX = PRR*log(10)/timePRR + GR.
}
\examples{
formula_PRRtoEMAX(0.1, 2)
formula_EMAXtoPRR(0.1, 0.195)

}
\seealso{
Other EMAX: 
\code{\link{formula_EC50toMPC90}()},
\code{\link{formula_EMAXEffect}()},
\code{\link{formula_EMAXmodelParsToMIC}()}

Other EMAX: 
\code{\link{formula_EC50toMPC90}()},
\code{\link{formula_EMAXEffect}()},
\code{\link{formula_EMAXmodelParsToMIC}()}
}
\author{
Anne Kuemmel (IntiQuan), Venelin Mitov (IntiQuan)
}
\concept{EMAX}
