% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{GeometricMean}
\alias{GeometricMean}
\title{Geometric mean of a positive numeric vector}
\usage{
GeometricMean(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector of positive values. If negative values are present,
a warning is issued and NA_real_ is returned.}

\item{na.rm}{logical indicating if NA values should be removed during the calculation.
If FALSE (the default) and NA values are present in x, then NA_real_ is returned.}
}
\value{
The geometric mean of x if all values ofor NA_real_if x contains non-positive values.
}
\description{
Geometric mean of a positive numeric vector
}
\examples{

x <- c(1, 2, 3, 4, 5)
GeometricMean(x)
GeometricMean(c(1, 2, 3, 4, 5, -1)) # issues a warning and returns NA_real_
GeometricMean(c(1, 2, 3, 4, 5, NA)) # returns NA_real_
GeometricMean(c(1, 2, 3, 4, 5, NA), na.rm = TRUE) # returns 2.605171
GeometricMean(c(1, 2, 3, 4, 5, NA, 0)) # issues a warning and returns NA_real_
GeometricMean(c(1, 2, 3, 4, 5, NA, 0), na.rm = TRUE) # issues a warning and returns NA_real_

}
\seealso{
Other Statistical: 
\code{\link{GeometricSD}()},
\code{\link{clopperPearsonMMV}()}
}
\author{
Venelin Mitov (IntiQuan)
}
\concept{Statistical}
