% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_ActivityPath}
\alias{get_ActivityPath}
\title{Get the directory path of an activity}
\usage{
get_ActivityPath(ActivityPath = NULL)
}
\arguments{
\item{ActivityPath}{If the user want to force the activity path to plot. (Default: \code{NULL}).}
}
\value{
A character vector with the activity path. Some examples to understand how the function
behaves:
\itemize{
\item If no ../tags.RData file exists:
\itemize{
\item current directory: /Projects_Discovery/Serie1/Work/CompX/A01_PK/Scripts - activity path: Serie1/Work/CompX/A01_PK
\item current directory: /Projects/CompX/Work/A01_PK/Scripts - activity path: CompX/Work/A01_PK
\item current directory: /some/other/path - activity path NULL with warning 'The current activity is not in the project folder'.
}
\item If file ../tags.RData exists and has entries: itemPath = "/sites/department/ModellingTeam/ProjectX",
itemName = "ActivityY", then ProjectX/ActivityY is returned.
}
}
\description{
Get activity path to be used in graphs. This allows to keep track of where the
figures are located and which script was used to generate them.
}
\details{
This function is used by some of the graphical functions in MMVbase. It is
not intended for use in scripts unless the user is working on an MMV activity.
}
\seealso{
Other Utility: 
\code{\link{aux_addNamesToPercentiles}()},
\code{\link{aux_formatErrorName}()},
\code{\link{aux_removeEscapeChar}()},
\code{\link{collapseMMV}()},
\code{\link{remove_duplicates}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org}),
Venelin Mitov (IntiQuan, \email{venelin.mitov@intiquan.com})
}
\concept{Utility}
