% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{get_fileNameExtension}
\alias{get_fileNameExtension}
\title{Get the extension of a filename}
\usage{
get_fileNameExtension(filename)
}
\arguments{
\item{filename}{Path to a file}
}
\value{
TRUE/FALSE
}
\description{
Get the extension of a filename
}
\seealso{
Other File: 
\code{\link{aux_CommonSubPath}()},
\code{\link{file.copyMMV}()},
\code{\link{is.fileMMV}()},
\code{\link{load_RDataAsList}()},
\code{\link{saveMMV}()},
\code{\link{thisFile}()}
}
\author{
Mohammed H. Cherkaoui (MMV), \email{cherkaouim@mmv.org}
}
\concept{File}
