% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{get_FromTable}
\alias{get_FromTable}
\title{Extract a value for a parameter or a constant from a data.frame such as a report table}
\usage{
get_FromTable(
  table,
  Parameter,
  Column = "Value",
  ValueIfNotFound = NULL,
  UnitIfNotFound = NULL,
  FLAG_USE_GREPL = FALSE,
  ...
)
}
\arguments{
\item{table}{a data.frame having a Parameter and, at least, a column named as the
\code{Column} parameter.}

\item{Parameter}{a character string denoting a Parameter name or a regular
expression to be searched for in \code{table$Parameter}.}

\item{Column}{a character string (default: "Value") indicating the column
for which the value to be returned once a row matching
\code{'Parameter'} is found.}

\item{ValueIfNotFound}{a value to return if Parameter not found in table.}

\item{UnitIfNotFound}{a character string to return if Parameter not found in table.}

\item{FLAG_USE_GREPL}{logical (default: FALSE) indicating if \code{\link{grepl}}
should be used instead of \code{\link{match}} for
finding}

\item{...}{additional arguments passed to \code{\link{grepl}} if FLAG_USE_GREPL is TRUE.}
}
\value{
If a matching row for Parameter was found, the Value with added attribute
'row' containing the matching row, otherwize \code{ValueIfNotFound} also with
added attribute 'row'.
}
\description{
Extract a value for a parameter or a constant from a data.frame such as a report table
}
\examples{
# Assume we have a report table called 'resultTable'
resultTable <- data.frame(
  Parameter = c("Selected Prediction Method", "Total Clearance", "Hepatic Clearance"),
  Human = c("IVIVE", 20, 17),
  Unit = c("method", "mL/min/kg", "mL/min/kg")
)

# Get the value for the 'Total Clearance' parameter for 'Human'
totalClearance <- get_FromTable(resultTable, "Total Clearance", "Human")
print(totalClearance)
}
\seealso{
Other Reporting: 
\code{\link{ConvertReportTableToList}()},
\code{\link{CreateOrAppendToReportTable}()},
\code{\link{PrintReportTableAsRmd}()}
}
\author{
Venelin Mitov
}
\concept{Reporting}
