% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericConstants.R
\name{get_NumericConstants}
\alias{get_NumericConstants}
\title{Get numeric constants such as growth rates, body weights, and cell numbers}
\usage{
get_NumericConstants()
}
\value{
a data.frame with Parameter, Value, Unit, and Source columns.
}
\description{
Get numeric constants such as growth rates, body weights, and cell numbers
}
\examples{

dfConstants <- get_NumericConstants()
dfConstants[
  startsWith(dfConstants$Name_unit, "BodyWeight"), c("Parameter", "Value", "Unit", "Source")]

}
