% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericConstants.R
\name{get_NumericConstantsAsList}
\alias{get_NumericConstantsAsList}
\title{Get numeric constants as a named list}
\usage{
get_NumericConstantsAsList()
}
\value{
a named list.
}
\description{
Get numeric constants as a named list
}
\examples{

get_NumericConstantsAsList()$MicrosomesDog_mgProteinPergLiver

}
