% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{IQRtableToDataFrame}
\alias{IQRtableToDataFrame}
\title{Convert saved IQR table back to data.frame}
\usage{
IQRtableToDataFrame(inputFileName)
}
\arguments{
\item{inputFileName}{Path to text file saved with \code{IQRtools::IQRoutputTable} to convert back to data.frame}
}
\description{
Convert saved IQR table back to data.frame
}
\seealso{
Other General Functions: 
\code{\link{aux_CommonSubPath}()},
\code{\link{aux_createUSUBJID}()},
\code{\link{aux_formatErrorName}()},
\code{\link{aux_removeEscapeChar}()},
\code{\link{clopperPearsonMMV}()},
\code{\link{convert_Unit}()},
\code{\link{dichotomicSearch}()},
\code{\link{file.copyMMV}()},
\code{\link{find_MinMMV}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rectintMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{thisFile}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
