% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{is.fileMMV}
\alias{is.fileMMV}
\title{Checks if filename is an existing file and not a directory}
\usage{
is.fileMMV(filename)
}
\arguments{
\item{filename}{Path to a file}
}
\value{
TRUE/FALSE
}
\description{
Checks if filename is an existing file and not a directory
}
\seealso{
\link[base:files]{base::file.exists}, \link[base:files2]{base::dir.exists}

Other File: 
\code{\link{aux_CommonSubPath}()},
\code{\link{file.copyMMV}()},
\code{\link{get_fileNameExtension}()},
\code{\link{load_RDataAsList}()},
\code{\link{saveMMV}()},
\code{\link{thisFile}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{File}
