% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsInInterval.R
\name{IsInOpenInterval}
\alias{IsInOpenInterval}
\title{Check if value is in the open interval \code{(low, high)}}
\usage{
IsInOpenInterval(value, low, high)
}
\arguments{
\item{value}{numeric.}

\item{low}{numeric.}

\item{high}{numeric.}
}
\value{
TRUE or FALSE. If any of the arguments is not numeric or low < high is not true an
error is raised.
}
\description{
Check if value is in the open interval \code{(low, high)}
}
\examples{
IsInClosedInterval(2, 1, 3)
IsInClosedInterval(2, 1, 2)
IsInClosedInterval(2, 1, 1.5)
error <- try(IsInClosedInterval(2, 1.5, 1), silent = TRUE)
}
\author{
Venelin Mitov (IntiQuan)
}
