% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical.R
\name{IsInRightOpenInterval}
\alias{IsInRightOpenInterval}
\title{Check if value is in the right open interval \code{[low, high)}}
\usage{
IsInRightOpenInterval(value, low, high)
}
\arguments{
\item{value}{numeric.}

\item{low}{numeric.}

\item{high}{numeric.}
}
\value{
TRUE or FALSE. If any of the arguments is not numeric or low < high is not true an
error is raised.
}
\description{
Check if value is in the right open interval \code{[low, high)}
}
\examples{
IsInRightOpenInterval(2, 1, 3)
IsInRightOpenInterval(2, 1, 2)
IsInRightOpenInterval(2, 1, 1.5)
error <- try(IsInRightOpenInterval(2, 1.5, 1), silent = TRUE)

}
\seealso{
Other Mathematical: 
\code{\link{IsInClosedInterval}()},
\code{\link{IsInLeftOpenInterval}()},
\code{\link{IsInOpenInterval}()},
\code{\link{find_MinMMV}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rectintMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{trapzMMV}()}
}
\author{
Venelin Mitov (IntiQuan)
}
\concept{Mathematical}
