% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{load_RDataAsList}
\alias{load_RDataAsList}
\title{Load RData into a list}
\usage{
load_RDataAsList(fileName)
}
\arguments{
\item{fileName}{Path to the RData file name.}
}
\value{
List with all object saved in \code{fileName}
}
\description{
Load RData into a list
}
\seealso{
Other File: 
\code{\link{aux_CommonSubPath}()},
\code{\link{file.copyMMV}()},
\code{\link{get_fileNameExtension}()},
\code{\link{is.fileMMV}()},
\code{\link{saveMMV}()},
\code{\link{thisFile}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{File}
