% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{logLinTrapzMMV}
\alias{logLinTrapzMMV}
\title{log-Linear Trapez Integration Method}
\usage{
logLinTrapzMMV(x, y, logCrieria = 0.99)
}
\arguments{
\item{x}{Numerical vector such as y=f(x)}

\item{y}{Numerical vector such as y=f(x)}

\item{logCrieria}{Criteria to use the log-lin rule (Default: 0.99)}
}
\description{
This function estimates the integral of a function using the trapeze method.
It uses the log-lin rule when the function is decreasing.
}
\seealso{
Other General Functions: 
\code{\link{dichotomicSearch}()},
\code{\link{rectintMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{thisFile}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
