% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical.R
\name{newtonRaphson.Function}
\alias{newtonRaphson.Function}
\title{Find the root of of y=f(x) within an interval}
\usage{
newtonRaphson.Function(
  funcToEval,
  Ysol = 0,
  x0 = NULL,
  range = c(-10, 10),
  relTol = 1e-06,
  absTol = 1e-09,
  n = 1000,
  iterMax = 1000
)
}
\arguments{
\item{funcToEval}{A function to evaluate for the equation.}

\item{Ysol}{A numeric value representing the solution such that the function returns the solution Xsol of Ysol = f(Xsol). Default is 0.}

\item{x0}{Initial value for the algorithm. If not provided, it will analyze the derivative to find the optimum start point. Default is \code{NULL}.}

\item{range}{A numeric vector of two scalars that determine the range of the search. Default is c(-10, 10).}

\item{relTol}{Relative tolerance of the solution. Default is 1e-06.}

\item{absTol}{Absolute tolerance of the solution. Default is 1e-09.}

\item{n}{Number of intervals to estimate the derivative. Default is 1000.}

\item{iterMax}{Maximum number of iterations. Default is 1000.}
}
\value{
A numeric value representing the root of the function.
}
\description{
This function uses the Newton-Raphson method to find the root of a given function.
}
\details{
The function uses the Newton-Raphson method to find the root of a given function. If the initial value \code{x0} is not provided, the function will analyze the derivative to find the optimum start point within the specified \code{range}. The algorithm iterates until the relative or absolute tolerance is met or the maximum number of iterations is reached.
}
\examples{
# Define a function
f <- function(x) { x^2 - 4 }
# Find the root using the Newton-Raphson method
newtonRaphson.Function(f)

}
\seealso{
Other Mathematical: 
\code{\link{IsInClosedInterval}()},
\code{\link{IsInLeftOpenInterval}()},
\code{\link{IsInOpenInterval}()},
\code{\link{IsInRightOpenInterval}()},
\code{\link{find_MinMMV}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rectintMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Mathematical}
