% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical.R
\name{newtonRaphson.Vector}
\alias{newtonRaphson.Vector}
\title{Newton-Raphson Method for Finding Roots in a Vector}
\usage{
newtonRaphson.Vector(
  X,
  Y,
  Ysol = 0,
  x0 = FALSE,
  relTol = 1e-06,
  absTol = 1e-09,
  eps = 1e-06,
  iterMax = 1000
)
}
\arguments{
\item{X}{A numeric vector of x-values where the function values are known.}

\item{Y}{A numeric vector of y-values corresponding to the function values at the x-values.}

\item{Ysol}{A numeric value representing the solution such that the function returns the solution Xsol of Ysol = f(Xsol). Default is 0.}

\item{x0}{Initial value for the algorithm. If not provided, it will analyze the derivative to find the optimum start point. Default is \code{FALSE}.}

\item{relTol}{Relative tolerance of the solution. Default is 1e-06.}

\item{absTol}{Absolute tolerance of the solution. Default is 1e-09.}

\item{eps}{Step to use to estimate the derivative. Default is 1e-06.}

\item{iterMax}{Maximum number of iterations. Default is 1000.}
}
\value{
A numeric value representing the root of the function.
}
\description{
This function uses the Newton-Raphson method to find the root of a given function represented by vectors of x and y values.
}
\details{
The function uses the Newton-Raphson method to find the root of a given function represented by vectors of x and y values. If the initial value \code{x0} is not provided, the function will analyze the derivative to find the optimum start point. The algorithm iterates until the relative or absolute tolerance is met or the maximum number of iterations is reached.
}
\examples{
# Define vectors X and Y
X <- c(0, 1, 2, 3, 4)
Y <- c(0, 1, 4, 9, 16)
# Find the root using the Newton-Raphson method
newtonRaphson.Vector(X, Y)

}
\seealso{
Other Mathematical: 
\code{\link{IsInClosedInterval}()},
\code{\link{IsInLeftOpenInterval}()},
\code{\link{IsInOpenInterval}()},
\code{\link{IsInRightOpenInterval}()},
\code{\link{find_MinMMV}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{rectintMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Mathematical}
