% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTable.R
\name{PrintReportTableAsRmd}
\alias{PrintReportTableAsRmd}
\title{Print a report table in rmakrdown format}
\usage{
PrintReportTableAsRmd(resultTable)
}
\arguments{
\item{reportTable}{a data.frame - a report table}
}
\value{
nothing. This function only produces output to the standard output device
}
\description{
Print a report table in rmakrdown format
}
