% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical.R
\name{rectintMMV}
\alias{rectintMMV}
\title{Calculate the integral of a function using the Riemann sum method.}
\usage{
rectintMMV(x, y, FLAGright = FALSE)
}
\arguments{
\item{x}{Vector x where the values of the function f(x) are known}

\item{y}{Vector y such as y=f(x)}

\item{FLAGright}{Logical; if \code{TRUE}, the right Riemann sum is used. If \code{FALSE},
the left Riemann sum is used. Default is \code{FALSE}.}
}
\value{
A numeric value representing the integral of the function over the given points.
}
\description{
Calculate the integral of a function using the Riemann sum method.
}
\details{
The function calculates the integral of a function using the Riemann sum method.
It requires that the vectors \code{x} and \code{y} have the same length and contain at least two
points. The user can choose between the right Riemann sum and the left Riemann sum by
setting the \code{FLAGright} parameter.
}
\examples{
x <- c(0, 1, 2, 3, 4)
y <- c(0, 1, 4, 9, 16)
rectintMMV(x, y)
rectintMMV(x, y, FLAGright = TRUE)
}
\seealso{
Other Mathematical: 
\code{\link{IsInClosedInterval}()},
\code{\link{IsInLeftOpenInterval}()},
\code{\link{IsInOpenInterval}()},
\code{\link{IsInRightOpenInterval}()},
\code{\link{find_MinMMV}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{simpsonMMV}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Mathematical}
