% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoad_RData.R
\name{saveMMV}
\alias{saveMMV}
\title{saveMMV
Save R objects}
\usage{
saveMMV(list, file)
}
\arguments{
\item{list}{A character vector containing the names of objects to be saved.}

\item{file}{The name of the file where the data will be saved.}
}
\description{
\link{saveMMV} writes an external representation of R objects to the specified file.
It uses the function \link[base:save]{base::save}. If 'file' is a path, it will create the directory
if not existent unlike \link[base:save]{base::save}.
}
\seealso{
\link[base:save]{base::save}

Other Save RData: 
\code{\link{load_RDataAsList}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Save RData}
