% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical.R
\name{simpsonMMV}
\alias{simpsonMMV}
\title{Simpson's Rule Integration}
\usage{
simpsonMMV(x, y, tolInt = 1e-09)
}
\arguments{
\item{x}{A numeric vector of x-values where the function values are known.}

\item{y}{A numeric vector of y-values corresponding to the function values at the x-values.}

\item{tolInt}{A numeric value specifying the tolerance to check if the step size is similar
among the intervals. Default is 1e-09.}
}
\value{
A numeric value representing the integral of the function over the given points.
}
\description{
This function performs numerical integration using Simpson's rule. If the number of
points is less than three, it falls back to the trapezoidal rule.
}
\details{
The function calculates the integral of a function using Simpson's rule. If the
number of points is less than three, it uses the trapezoidal rule. It requires that the
vectors \code{x} and \code{y} have the same length and contain at least two points. If the step sizes
are not consistent, it adapts the Simpson's rule accordingly.
}
\examples{
x <- c(0, 1, 2, 3, 4)
y <- c(0, 1, 4, 9, 16)
simpsonMMV(x, y)

}
\seealso{
Other Mathematical: 
\code{\link{IsInClosedInterval}()},
\code{\link{IsInLeftOpenInterval}()},
\code{\link{IsInOpenInterval}()},
\code{\link{IsInRightOpenInterval}()},
\code{\link{find_MinMMV}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rectintMMV}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Mathematical}
