% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{summaryAcrossTrials}
\alias{summaryAcrossTrials}
\title{Summarize confidence intervals for provided metrics across trials.}
\usage{
summaryAcrossTrials(
  dataInput,
  CIlevel = 90,
  metricCOL = "Metric",
  variableCOL = "Variable",
  valueCOL = "Value",
  scenCOL = c("ScenID", "ExpID", "DoseID", "Dose", "nbrDoses", "TIME")
)
}
\arguments{
\item{dataInput}{data.table containing summarized outcome variable(s) summarized by trials - produced by \code{summaryByTrial} or similar format.}

\item{CIlevel}{Numeric containing the confidence interval in percent to be estimated (Default: 90)}

\item{metricCOL}{name of column of \code{dataInput} containing names of metrics on which CI is being estimated (Default: \code{"Metric"})}

\item{variableCOL}{name of column of \code{dataInput} containing names of variables (Default: \code{"Variable"})}

\item{valueCOL}{name of column of \code{dataInput} containing values of metrics for specific variables (Default: \code{"Value"})}

\item{scenCOL}{Name(s) of column(s) of \code{dataInput} uniquely describing scenario (Default: \code{c("ScenID","ExpID","DoseID","Dose","nbrDoses","TIME")})}
}
\value{
data table containing confidence intervals for selected outcome variable(s) and summarized metric(s)
}
\description{
Summarize confidence intervals for provided metrics across trials.
}
\seealso{
Other Data: 
\code{\link{Check_MissingDatabyNAME}()},
\code{\link{IQRtableToDataFrame}()},
\code{\link{aux_constructCIpercentiles}()},
\code{\link{aux_createUSUBJID}()},
\code{\link{check_dataGeneralMMV}()},
\code{\link{check_dataGeneralMMV_Center}()},
\code{\link{convert_Unit}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{summarize_PIandCIgeneric}()},
\code{\link{summaryByTrial}()},
\code{\link{swapName_MMVnameToName}()},
\code{\link{swapName_NameToMMVname}()},
\code{\link{transform_dataFrame_LongToWide}()},
\code{\link{transform_dataFrame_WideToLong}()}
}
\author{
Sam Jones (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{Data}
