% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{transform_dataFrame_WideToLong}
\alias{transform_dataFrame_WideToLong}
\title{transform_dataFrame}
\usage{
transform_dataFrame_WideToLong(
  dataFrame_Wide,
  key,
  value,
  colMaster,
  colPattern = paste0(".", value)
)
}
\arguments{
\item{colPattern}{Default: paste0(".", value)}
}
\description{
transform_dataFrame
}
\seealso{
Other Data Preparation: 
\code{\link{Check_MissingDatabyNAME}()},
\code{\link{check_dataGeneralMMV}()},
\code{\link{check_dataGeneralMMV_Center}()},
\code{\link{swapName_MMVnameToName}()},
\code{\link{swapName_NameToMMVname}()},
\code{\link{transform_dataFrame_LongToWide}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Data Preparation}
