% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical.R
\name{transform_IQRggplotToMMVggplot}
\alias{transform_IQRggplotToMMVggplot}
\title{Transform IQR ggplot object to MMV ggplot object}
\usage{
transform_IQRggplotToMMVggplot(
  IQRggplot,
  style = c("MMV", "MMVvpc")[1],
  ActivityPath = NULL,
  Caption = NULL
)
}
\arguments{
\item{IQRggplot}{ggplot object generated with \code{IQRtools} functions.}

\item{style}{Style to choose between \code{"MMV"} or \code{"MMVvpc"}
(Default: \code{"MMV"}).}

\item{ActivityPath}{Path of the current activity (Default: \code{NULL})}

\item{Caption}{Caption to add to the plot (Default: \code{NULL})}
}
\value{
A ggplot object compliant with MMV standars.
}
\description{
When a ggplot object is generated using \code{IQRtools} functions, this function
allows to adjust the ggplot object to be compliant with MMV standards;
It will add the path of the activity and adjust the theme if specified.
}
\seealso{
Other Graphical: 
\code{\link{MMVcolors}},
\code{\link{MMVggplot}()},
\code{\link{MMVtheme}()},
\code{\link{adjust_ggplotTheme}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org}),
Venelin Mitov (IntiQuan, \email{venelin.mitov@intiquan.com})
}
\concept{Graphical}
