% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{trapzMMV}
\alias{trapzMMV}
\title{trapzMMV}
\usage{
trapzMMV(x, y)
}
\description{
trapzMMV
}
\seealso{
Other General Functions: 
\code{\link{dichotomicSearch}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{rectintMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{thisFile}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
