% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{aux_constructCIpercentiles}
\alias{aux_constructCIpercentiles}
\title{Construct CI percentiles from CI level}
\usage{
aux_constructCIpercentiles(CIlevel)
}
\arguments{
\item{CIlevel}{Numeric containing the confidence interval in percent to be estimated (Default: 90)}
}
\value{
The vector \code{CI.percentiles} with names.
}
\description{
Construct CI percentiles from CI level
}
\seealso{
Other Data: 
\code{\link{Check_MissingDatabyNAME}()},
\code{\link{IQRtableToDataFrame}()},
\code{\link{aux_createUSUBJID}()},
\code{\link{check_dataGeneralMMV_Center}()},
\code{\link{check_dataGeneralMMV}()},
\code{\link{convert_Unit}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{summarize_PIandCIgeneric}()},
\code{\link{summaryAcrossTrials}()},
\code{\link{summaryByTrial}()},
\code{\link{swapName_MMVnameToName}()},
\code{\link{swapName_NameToMMVname}()},
\code{\link{transform_dataFrame_LongToWide}()},
\code{\link{transform_dataFrame_WideToLong}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Data}
