% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{aux_createUSUBJID}
\alias{aux_createUSUBJID}
\title{Create USUBJID}
\usage{
aux_createUSUBJID(data)
}
\arguments{
\item{data}{Dataset for which to create \code{USUBJID}}
}
\value{
\code{data} with extra column \code{USUBJID}
}
\description{
Add column \code{USUBJID} to \code{data} based on columns \code{STUDY}, \code{GROUP} and \code{SUBJECT},
when preparing IQR data from SCID dataset
}
\examples{
data <- data.frame(STUDY   = "StudyID",
                   GROUP   = c("G1","G1","G2","G2"),
                   SUBJECT = c("M1","M2", "M1", "M2"),
                   stringsAsFactors = FALSE)
data$USUBJID <- MMVmalaria:::aux_createUSUBJID(data)

}
\seealso{
Other General Functions: 
\code{\link{IQRtableToDataFrame}()},
\code{\link{aux_CommonSubPath}()},
\code{\link{aux_formatErrorName}()},
\code{\link{aux_removeEscapeChar}()},
\code{\link{clopperPearsonMMV}()},
\code{\link{convert_Unit}()},
\code{\link{dichotomicSearch}()},
\code{\link{file.copyMMV}()},
\code{\link{find_MinMMV}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rectintMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{thisFile}()},
\code{\link{trapzMMV}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan)
}
\concept{General Functions}
