% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{aux_formatErrorName}
\alias{aux_formatErrorName}
\title{Format Error Name}
\usage{
aux_formatErrorName(parameterNames)
}
\arguments{
\item{parameterNames}{Vector of character with the names of the parameters}
}
\value{
\code{parameterNames} with converted format
}
\description{
Converts Error Parameter Names as used in sysFIT to format as used in NLME
}
\seealso{
Other General Functions: 
\code{\link{IQRtableToDataFrame}()},
\code{\link{aux_CommonSubPath}()},
\code{\link{aux_createUSUBJID}()},
\code{\link{aux_removeEscapeChar}()},
\code{\link{clopperPearsonMMV}()},
\code{\link{convert_Unit}()},
\code{\link{dichotomicSearch}()},
\code{\link{file.copyMMV}()},
\code{\link{find_MinMMV}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rectintMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{thisFile}()},
\code{\link{trapzMMV}()}
}
\author{
Anne Kuemmel (IntiQuan)
}
\concept{General Functions}
