% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{check_dataGeneralMMV}
\alias{check_dataGeneralMMV}
\title{check_dataGeneralMMV
Checks if CENTER and CENTERNAME is well defined within \code{dataGeneral}}
\usage{
check_dataGeneralMMV(dataGeneral)
}
\arguments{
\item{dataGeneral}{IQRdataGENERAL object}
}
\value{
IQRdataGENERAL object
}
\description{
Checks if CENTER and CENTERNAME (number and name of the center conducting the study) is defined within \code{dataGeneral}, according to MMV's standards.
}
\details{
The name and number of the center conducting the study will be checked according to the up-to-date center list, see \link{list_Center}.
The \link{check_dataGenetalMMV_Center} function is used internally.
}
\seealso{
\link{list_Center}, \link{check_dataGenetalMMV_Center}

Other Data Preparation: 
\code{\link{Check_MissingDatabyNAME}()},
\code{\link{check_dataGeneralMMV_Center}()}
}
\author{
Mohammed H. Cherkaoui (MMV), \email{cherkaouim@mmv.org}
}
\concept{Data Preparation}
