% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{clopperPearsonMMV}
\alias{clopperPearsonMMV}
\title{Clopper-Pearson Test}
\usage{
clopperPearsonMMV(k, n, CI = 0.9)
}
\arguments{
\item{k}{Number of successful observations of a trial}

\item{n}{Number of observation}

\item{CI}{Confidence Interval to estimate (Default: \code{0.9})}
}
\value{
Return a vector with the lower \code{CI_low} and high \code{CI_high} limit of the CI.
}
\description{
Statistical test to estimate the confidence interval of
a binomial distribution using the Clopper-Pearson Test
(see \href{https://en.wikipedia.org/wiki/Binomial_proportion_confidence_interval#Clopper\%E2\%80\%93Pearson_interval}{Wikipedia})
}
\examples{
# Probability of success:
p <- 0.8
# Number of observations:
n <- 100
# Number of successful observation:
k <- rbinom(1,n,p)
# Empirical probability of success:
p_Empirical <- k/n
# 90\% confidence interval of the empirical probability of success:
CI <- clopperPearsonMMV(k,n,CI=0.9)

}
\seealso{
Other Statistical: 
\code{\link{GeometricMean}()},
\code{\link{GeometricSD}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Statistical}
