% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{ConvertReportTableToList}
\alias{ConvertReportTableToList}
\title{Convert a reportTable to a list of data.frames corresponding to non-NA
entries in the Style column}
\usage{
ConvertReportTableToList(table)
}
\arguments{
\item{table}{a data.frame with a character column named Style.}
}
\value{
a list of data.tables
}
\description{
Convert a reportTable to a list of data.frames corresponding to non-NA
entries in the Style column
}
\seealso{
Other Reporting: 
\code{\link{CreateOrAppendToReportTable}()},
\code{\link{PrintReportTableAsRmd}()},
\code{\link{get_FromTable}()}
}
\author{
Venelin Mitov
}
\concept{Reporting}
