% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translational.R
\name{formula_CLintuToCLhPl}
\alias{formula_CLintuToCLhPl}
\title{Estimate hepatic plasma CL, based on intrinsic unbound hepatic clearance, hepatic blood flow,
fraction unbound in plasma and BP-ratio}
\usage{
formula_CLintuToCLhPl(CLintu, Qh, Fu, BP)
}
\arguments{
\item{CLintu}{intrinsic unbound hepatic clearance (mL/min/kg).}

\item{Qh}{hepatic blood flow (mL/min/kg).}

\item{Fu}{fraction unbound in plasma (.).}

\item{BP}{BP-ratio (.).}
}
\value{
Estimated hepatic plasma clearance (mL/min/kg).
}
\description{
Estimate hepatic plasma CL, based on intrinsic unbound hepatic clearance, hepatic blood flow,
fraction unbound in plasma and BP-ratio
}
\details{
Currently using the well-stirred model.
}
\seealso{
Other Translational: 
\code{\link{formula_CLhPlToCLintu}()},
\code{\link{formula_CLhToCLintu}()},
\code{\link{formula_CLintuToCLh}()},
\code{\link{formula_CLintuToCLh_Parallel}()},
\code{\link{formula_CLintuToCLh_WellStirred}()},
\code{\link{formula_HepExtraction}()},
\code{\link{formula_convertConc}()}
}
\author{
Venelin Mitov (IntiQuan)
}
\concept{Translational}
