% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_ActivityPath}
\alias{get_ActivityPath}
\title{Get Activity Path}
\usage{
get_ActivityPath(ActivityPath = NULL)
}
\arguments{
\item{ActivityPath}{If the user want to force the activity path to plot. (Default: \code{NULL}).}
}
\value{
A character vector with the activity path
}
\description{
Get activity path to be used in graphs. This allows to keep track of where the
figures are located and which script was used to generate them.
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org}),
Venelin Mitov (IntiQuan, \email{venelin.mitov@intiquan.com})
}
