% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical.R
\name{IsInClosedInterval}
\alias{IsInClosedInterval}
\title{Check if value is in the closed interval \code{[low, high]}}
\usage{
IsInClosedInterval(value, low, high)
}
\arguments{
\item{value}{numeric.}

\item{low}{numeric.}

\item{high}{numeric.}
}
\value{
TRUE or FALSE. If any of the arguments is not numeric or low <= high is not true an
error is raised.
}
\description{
Check if value is in the closed interval \code{[low, high]}
}
\examples{
IsInClosedInterval(2, 1, 3)
IsInClosedInterval(2, 1, 2)
IsInClosedInterval(2, 1, 1.5)

}
\seealso{
Other Mathematical: 
\code{\link{IsInLeftOpenInterval}()},
\code{\link{IsInOpenInterval}()},
\code{\link{IsInRightOpenInterval}()},
\code{\link{find_MinMMV}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rectintMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{trapzMMV}()}
}
\author{
Venelin Mitov (IntiQuan)
}
\concept{Mathematical}
