% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsInInterval.R
\name{IsInLeftOpenInterval}
\alias{IsInLeftOpenInterval}
\title{Check if value is in the left open interval \code{(low, high]}}
\usage{
IsInLeftOpenInterval(value, low, high)
}
\arguments{
\item{value}{numeric.}

\item{low}{numeric.}

\item{high}{numeric.}
}
\value{
TRUE or FALSE. If any of the arguments is not numeric or low < high is not true an
error is raised.
}
\description{
Check if value is in the left open interval \code{(low, high]}
}
\examples{
IsInLeftOpenInterval(2, 1, 3)
IsInLeftOpenInterval(2, 1, 2)
IsInLeftOpenInterval(2, 1, 1.5)
error <- try(IsInLeftOpenInterval(2, 1.5, 1), silent = TRUE)
}
\author{
Venelin Mitov (IntiQuan)
}
