% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoad_RData.R
\name{load_RDataAsList}
\alias{load_RDataAsList}
\title{Load RData into a list}
\usage{
load_RDataAsList(fileName)
}
\arguments{
\item{fileName}{Path to the RData file name.}
}
\value{
List with all object saved in \code{fileName}
}
\description{
Load RData into a list
}
\seealso{
Other Save RData: 
\code{\link{saveMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Save RData}
