% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical.R
\name{logLinTrapzMMV}
\alias{logLinTrapzMMV}
\title{log-Linear Trapez Integration Method}
\usage{
logLinTrapzMMV(x, y, logCrieria = 0.99)
}
\arguments{
\item{x}{Numerical vector such as y=f(x)}

\item{y}{Numerical vector such as y=f(x)}

\item{logCrieria}{Criteria to use the log-lin rule (Default: 0.99)}
}
\value{
A numeric value representing the integral of the function over the given points.
}
\description{
This function estimates the integral of a function using the trapeze method.
It uses the log-lin rule when the function is decreasing.
}
\details{
The function calculates the integral of a function using the trapezoidal rule.
If the function is decreasing and the y-values meet the logCrieria, it uses the log-linear
rule for integration. It requires that the vectors \code{x} and \code{y} have the same length and
contain at least two points.
}
\examples{
x <- c(1, 2, 3, 4, 5)
y <- c(10, 8, 6, 4, 2)
logLinTrapzMMV(x, y)

}
\seealso{
Other Mathematical: 
\code{\link{IsInClosedInterval}()},
\code{\link{IsInLeftOpenInterval}()},
\code{\link{IsInOpenInterval}()},
\code{\link{IsInRightOpenInterval}()},
\code{\link{find_MinMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rectintMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Mathematical}
