% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\docType{data}
\name{MMVcolors}
\alias{MMVcolors}
\title{Color table for MMVbase}
\format{
An object of class \code{character} of length 5000.
}
\usage{
MMVcolors
}
\description{
Color table for MMVbase
}
\author{
Venelin Mitov (IntiQuan, \email{venelin.mitov@intiquan.com})
}
\keyword{internal}
