% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{MMVtheme}
\alias{MMVtheme}
\title{MMVbase theme for ggplot graphics}
\usage{
MMVtheme(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{numeric, font-size (default:12).}

\item{base_family}{character, font-name (default: "").}
}
\value{
a ggplot2 theme object.
}
\description{
The standard gg theme is based on theme_bw.
}
\author{
Venelin Mitov (venelin.mitov@intiquan.com)
}
