% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{PrintReportTableAsRmd}
\alias{PrintReportTableAsRmd}
\title{Print a report table in rmakrdown format}
\usage{
PrintReportTableAsRmd(resultTable, colnames = NULL)
}
\arguments{
\item{resultTable}{a data.frame or data.table as returned by CreateOrAppendToReportTable.}

\item{colnames}{a character vector (default NULL) used to reorder or select some of the columns in the
table to be rendered as rmd. If NULL all columns after IncludeInReport are printed in their order as in
the resultTable.}
}
\value{
nothing. This function only produces output to the standard output device
}
\description{
Print a report table in rmakrdown format
}
\seealso{
Other Reporting: 
\code{\link{ConvertReportTableToList}()},
\code{\link{CreateOrAppendToReportTable}()},
\code{\link{get_FromTable}()}
}
\concept{Reporting}
