% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{rectintMMV}
\alias{rectintMMV}
\title{Calculate the integral of a function using the Riemann sum method.}
\usage{
rectintMMV(x, y, FLAGright = FALSE)
}
\arguments{
\item{x}{Vector x where the values of the function f(x) are known}

\item{y}{Vector y such as y=f(x)}

\item{FLAGright}{Default: FALSE}
}
\value{
The integral of the function f(x) using the Riemann sum method.
}
\description{
Calculate the integral of a function using the Riemann sum method.
}
\seealso{
Other General Functions: 
\code{\link{dichotomicSearch}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{thisFile}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
