% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{simpsonMMV}
\alias{simpsonMMV}
\title{simpsonMMV}
\usage{
simpsonMMV(x, y, tolInt = 1e-09)
}
\arguments{
\item{tolInt}{Default: 1e-09}
}
\description{
simpsonMMV
}
\seealso{
Other General Functions: 
\code{\link{dichotomicSearch}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{rectintMMV}()},
\code{\link{thisFile}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
