% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{simpsonMMV}
\alias{simpsonMMV}
\title{simpsonMMV}
\usage{
simpsonMMV(x, y, tolInt = 1e-09)
}
\arguments{
\item{tolInt}{Default: 1e-09}
}
\description{
simpsonMMV
}
\seealso{
Other General Functions: 
\code{\link{IQRtableToDataFrame}()},
\code{\link{aux_CommonSubPath}()},
\code{\link{aux_createUSUBJID}()},
\code{\link{aux_formatErrorName}()},
\code{\link{aux_removeEscapeChar}()},
\code{\link{clopperPearsonMMV}()},
\code{\link{convert_Unit}()},
\code{\link{dichotomicSearch}()},
\code{\link{file.copyMMV}()},
\code{\link{find_MinMMV}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rectintMMV}()},
\code{\link{thisFile}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
