% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summaryByTrial}
\alias{summaryByTrial}
\title{Summarize given outcome variable(s) for a single trial}
\usage{
summaryByTrial(
  dataInput,
  varCOL,
  percentiles = c(5, 50, 95),
  usubjidCOL = c("IndivID", "USUBJID"),
  trialCOL = c("ScenID", "ExpID", "DoseID", "Dose", "nbrDoses", "TrialID", "TIME")
)
}
\arguments{
\item{dataInput}{data.table containing outcome variable(s) and selected other variables for each individual.}

\item{varCOL}{List containing metrics (percentiles, mean and/or geomean) to be calculated for each outcome variable - i.e. \code{list(Cmax=c("Percentile","Mean"),APR="Mean")}.}

\item{percentiles}{Percentiles of outcome variable(s) to be calculated if appropriate (Default: \code{c(5,50,90)}).}

\item{trialCOL}{Name(s) of column(s) of \code{dataInput} uniquely describing trial (Default: \code{c("ScenID","ExpID","DoseID","Dose","nbrDoses","TrialID","TIME")}).}

\item{usubJIDCOL}{Name(s) of column(s) of \code{dataInput} uniquely describing individuals (Default: \code{c("IndivID","USUBJID")}).}
}
\value{
data table containing values of outcome metrics for each trial
}
\description{
Summarize given outcome variable(s) for a single trial using a choice of metrics
(mean, percentile or geomean). If no metric is provided, percentile is used by default
for numeric variable, while mean is used for logical variable.
}
\seealso{
Other Summarize Functions: 
\code{\link{summarize_PIandCIgeneric}()},
\code{\link{summaryAcrossTrials}()}
}
\author{
Sam Jones (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{Summarize Functions}
