% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{swapName_MMVnameToName}
\alias{swapName_MMVnameToName}
\title{swapName_MMVnameToName
Swaps from MMV name for compound for nick name of interest}
\usage{
swapName_MMVnameToName(data, CompoundList)
}
\arguments{
\item{data}{IQRdataGENERAL object, data frame or character vector with MMV name for compound}

\item{CompoundList}{list with both MMV and nick name for the compound}
}
\value{
\code{data} object with nick name for compound
}
\description{
Swaps the MMV name for the nick name of interest within \code{data} based on \code{CompoundList}.
If \code{data} is a IQRdataGENERAL object, MMV name for compound will also be swapped in IQRdataGENERAL object attributes.
}
\seealso{
Other Data Preparation: 
\code{\link{Check_MissingDatabyNAME}()},
\code{\link{check_dataGeneralMMV}()},
\code{\link{check_dataGeneralMMV_Center}()},
\code{\link{swapName_NameToMMVname}()},
\code{\link{transform_dataFrame_LongToWide}()},
\code{\link{transform_dataFrame_WideToLong}()}
}
\author{
To be defined
}
\concept{Data Preparation}
