% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thisFile.R
\name{thisFile}
\alias{thisFile}
\title{thisFile}
\usage{
thisFile(fullName = FALSE, pathNormalized = FALSE)
}
\arguments{
\item{fullName}{Default: FALSE}

\item{pathNormalized}{Default: FALSE}
}
\description{
thisFile
}
\seealso{
Other General Functions: 
\code{\link{dichotomicSearch}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{rectintMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
