% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{thisFile}
\alias{thisFile}
\title{thisFile}
\usage{
thisFile(fullName = FALSE, pathNormalized = FALSE)
}
\arguments{
\item{fullName}{Default: FALSE}

\item{pathNormalized}{Default: FALSE}
}
\description{
thisFile
}
\seealso{
Other General Functions: 
\code{\link{IQRtableToDataFrame}()},
\code{\link{aux_CommonSubPath}()},
\code{\link{aux_createUSUBJID}()},
\code{\link{aux_formatErrorName}()},
\code{\link{aux_removeEscapeChar}()},
\code{\link{clopperPearsonMMV}()},
\code{\link{convert_Unit}()},
\code{\link{dichotomicSearch}()},
\code{\link{file.copyMMV}()},
\code{\link{find_MinMMV}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rectintMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
