% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{transform_dataFrame_LongToWide}
\alias{transform_dataFrame_LongToWide}
\title{transform_dataFrame_LongToWide}
\usage{
transform_dataFrame_LongToWide(
  dataFrame_Long,
  key,
  value,
  colMaster,
  patternConst = function(x) {
     out <- paste0(".", x)
 }
)
}
\arguments{
\item{patternConst}{Default: function(x) {
out <- paste0(".", x)
}}
}
\description{
transform_dataFrame_LongToWide
}
\seealso{
Other Data Preparation: 
\code{\link{Check_MissingDatabyNAME}()},
\code{\link{check_dataGeneralMMV}()},
\code{\link{check_dataGeneralMMV_Center}()},
\code{\link{swapName_MMVnameToName}()},
\code{\link{swapName_NameToMMVname}()},
\code{\link{transform_dataFrame_WideToLong}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Data Preparation}
