% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical.R
\name{trapzMMV}
\alias{trapzMMV}
\title{Trapezoidal Integration}
\usage{
trapzMMV(x, y)
}
\arguments{
\item{x}{A numeric vector of x-values where the function values are known.}

\item{y}{A numeric vector of y-values corresponding to the function values at the x-values.}
}
\value{
A numeric value representing the integral of the function over the given points.
}
\description{
This function performs numerical integration using the trapezoidal rule.
}
\details{
The function calculates the integral of a function using the trapezoidal rule.
It requires that the vectors \code{x} and \code{y} have the same length and contain at least two points.
}
\examples{
x <- c(0, 1, 2, 3, 4)
y <- c(0, 1, 4, 9, 16)
trapzMMV(x, y)

}
\seealso{
Other Mathematical: 
\code{\link{IsInClosedInterval}()},
\code{\link{IsInLeftOpenInterval}()},
\code{\link{IsInOpenInterval}()},
\code{\link{IsInRightOpenInterval}()},
\code{\link{find_MinMMV}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rectintMMV}()},
\code{\link{simpsonMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Mathematical}
